# pnetshieldq/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "version": "2.0"
    },
    "paths": {
        "/file/delete": {
            "get": {
                "description": "Delete a file from the server",
                "summary": "Delete a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The name of the file to delete",
                        "name": "fileName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/file/download": {
            "get": {
                "description": "Download a file by ID",
                "produces": [
                    "application/octet-stream"
                ],
                "summary": "Download a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the file to download",
                        "name": "fileID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File downloaded successfully"
                    },
                    "404": {
                        "description": "File not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/file/upload": {
            "post": {
                "description": "Upload a file to the server",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Upload a file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "The file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File uploaded successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/file/userfiles": {
            "get": {
                "description": "Get all files related to the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get user's files",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.File"
                            }
                        }
                    }
                }
            }
        },
        "/message/history/{userID}/{friendID}": {
            "get": {
                "description": "Get all messages between two users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get historical messages between two users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Friend ID",
                        "name": "friendID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/message/send": {
            "post": {
                "description": "Send a message to a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Send a message",
                "parameters": [
                    {
                        "description": "Message request",
                        "name": "messageRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.MessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Message"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "put": {
                "description": "update user by id",
                "summary": "Update User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete user by id",
                "summary": "Delete User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.MessageRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isHistorical": {
                    "type": "boolean"
                },
                "receiverId": {
                    "type": "integer"
                },
                "senderId": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "entity.File": {
            "type": "object",
            "properties": {
                "fileID": {
                    "type": "integer"
                },
                "fileName": {
                    "type": "string"
                },
                "fileType": {
                    "type": "string"
                },
                "fileURL": {
                    "type": "string"
                },
                "receiverID": {
                    "type": "string"
                },
                "senderID": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "entity.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "message_id": {
                    "type": "integer"
                },
                "message_type": {
                    "type": "string"
                },
                "receiver_id": {
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "entity.User": {
            "type": "object",
            "properties": {
                "clientID": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "lastSeen": {
                    "type": "string"
                },
                "onlineStatus": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}